Origin_/requirements.txt
torch
torchvision
transformers
pycocotools
Pillow
tqdm
opencv-python
scikit-learn
matplotlib
evaluate
